Add back in nearest road for Ahead of Me
I'd missed this by mistake when refactoring - this change restores the 'X
road ahead' prior to describing an intersection. However, it does bring up
a question:
  In the previous commit I switched nearestRoad to be the actual nearest
road, and not the nearest within the FOV. For reporting 'X road ahead' I
think that we do want the road in the FOV so I calculate both. It seems
'no worse', but requires some more thought. However, the code is all in one
place now so should be easier to consider.