When deduplicating point and Polygon POI allow for multiple Polygons
We don't want to have point POIs if a polygon POI can be creates instead.
However, often there are multiple polygons for a single POI and this change
ensures that we deal with that by maintaining a list of polygons rather
than just a single Polygon when checking for duplicates.