Adjust coordinate calculations for MVT translation by 0.5 pixel
The calculations that translate the integer relative tile coordinates into
fractional tile coordinates now place the sample point in the middle of the
sample range i.e. on the half pixel. This fixes the calculations made for the
interpolated points where lines cross tile boundaries as the tile boundaries
are no longer coincident with the sample points.