Add timeout to direction provider to update audio engine
When the phone is still, or when running in the emulator, the orientation
is not being updated as it's not changing. This was resulting in the audio
engine not being updated which was causing its state machine to stall.
Ideally that will be fixed inside the C++ code, but it's conceptually much
simpler to change the calling code here. If there's no heading update within
the timeout period of 100ms, then the audio engine is still called with the
last heading and the current location. Note that when running on the emulator
the location can change whilst the heading remains forever stuck at 0.0.