Rework FOV triangle calculations and extend scope of UserGeometry
The FOV triangle calculations were a bit beyond me, so here they are
reworked to share more common code. I've pinched the comments from the
iOS code to describe the different FOV modes.
Whilst doing that UserGeometry crept further into the code, so there
are a lot of changes to the unit tests to add that in.