Move Feature classification functions into GridState
This is the only place in the main app that calls the feature classification
code so it makes sense to be in that class. The main fallout is switching more
unit tests to use GridState so that they can still get hold of the
classifications. An interesting piece of fallout from this is that the
FeatureTree code wasn't dealing with MultiPolygons and so some unit tests
failed now that they were roundtripping FeatureCollections through the
FeatureTree. Adding support for MultiPolygon into FeatureTree fixes those
tests.