Refactor whatsAroundMe
The code that calculates selected super categories was being run each time
whatsAroundMe was called. This commit adds some new FeatureTrees for the
different super categories including one that is for those that are
currently selected. This simplifies the caller code significantly and moves
the cost of calculation to the tile update for a one off cost.

By using the FeatureTrees to search the super categories, we can remove
findFeaturesInPolygons which is the same code as already is in the
FeatureTree.

A couple of the distance functions have been moved into LatLngAlt for the
same reason as the original distance function was moved there. It makes
for a cleaner API for calling that has a less confusing argument list.