Add nearest road filter to better track the current road
This change adds a simple algorithm which calculates the nearest road on every
location update. It uses the current direction of travel as well as the current
location along with a bit of stickiness to the previous nearest road. The main
issue this aims to resolve is calling out the wrong nearest road when crossing
a side road, and also to ensure that intersections are called out from the correct
perspective.
The amount of Kalman filtering on the location has been increased slightly. This
increases the location lag, but with the benefit of more accurate location.