Attempt at tidying up routes navigation
There were a few crashes due to HomeRoutes being called on the routes
screen navigator. This tidies that up and everything is back working,
though the routes/markers selection doesn't persist when going back to
the home screen and back into the route screen. Not sure if that should
be expected?