Simplify markers/routes tab navigation
The desired behaviour is that the chosen tab is remembered, and swiping
back treats the Markers and Routes screen as a single screen. To be honest
I didn't fully understand the code before, and so I may have missed some
subtlety, but the way it works now is that the HomeViewModel stores the
selected tab and there's only a single NavController involved. As far as I
can tell this gives us the behaviour that we want.