Finessing of auto callout
More work here to include some of the iOS app logic. UserGeometry is no
longer just a data class, but includes functions to request various
distances which are calculated from it's member variables. The main
different with iOS auto callouts, is that on iOS lots of the callouts
get discarded because there's already speech being played.