Add places and landmarks tree and refine AroundMe
The AroundMe callout generator was searching amonst too many super
categories. This change limits it to just places and landmarks by creating
yet another tree to search in. Once we have all the functionality that we
need, we can remove unused trees.
The other change is that in whatsAroundMe the code now requests more than
1 features within each quadrant. This is so that de-duplication between
quadrants can be done properly.