Add some flow control for text-to-speech and callouts
The immediately useful change here is that the auto callout code now calls
isAudioEngineBusy before generating any more output. If there is already
text-to-speech playing out, then it won't generate any more. This aims to
prevent a backup of audio description.
The other change is the start of a callback when audio playout completes,
but so far this is only within the C++ code and doesn't make it back to the
kotlin code.