Initial decode of iOS exported route and import into database
This is based on a sample file from when a route was exported from the
iOS app. This commit registers application/json for the app and then
tries to decode it with the assumption that it matches the format in the
sample file. RouteData and RoutePoint don't have fields for all of those
in the JSON, though we may add them in future.

Sharing a marker from iOS results in a deep-link URL. Once we have a
server set up we can uncomment the support for it and add parsing for
that too.