Initial help topics screen from main menu
Code to display the help text originating from the iOS resources. The
Sections data structure is designed to make it easy to add in new
Sections of help, FAQ and tips.

The text for the resources has not been checked relative to how the
app actually works, so this is just the first step in their addition.