Initial addition of travel heading
We need similar behaviour to the iOS app where it sometimes uses the phone
heading, and sometimes the direction of travel as its heading. For example,
auto callouts always use the direction of travel, but triggered callouts like
AroundMe use the phone heading. This commit extends UserGeometry so that it
stores both types of heading and selects between them.