UI additions to allow waypoints to be added to routes and reordered
This change allows the selection of Markers to be added to Routes. It's
not refined from a UI point of view, but I think the code is at least
reasonably structured. Amonst the highlights:

* ReorderableLocationList adds a list of LocationItems which can be
dragged to reorder. Currently the whole item can be clicked to drag, but
really we just want a drag handle target on the end.

* Addition of ObjectIds to MarkerData and RouteData for unique identification.
This means that we can have multiple routes and markers with the same name
which matches iOS.

* Add `decoration` to `LocationItem` to allow its reuse in different list.

* Switch to using Flows from the Realm RouteRepository to enable lists to
dynamically update when the contents of the database changes. Fix the sorting
code so that it's closer to doing the right thing...

* Merge AddRoute and EditRoute screens into a single screen.