Earcon audio positioning fixes
The Earcon code was a bit confused and resulted in audio positioning not
working. This change makes EarconSource operate in the same was as
BeaconBuffer and TextToSpeechSource. It creates the sound in FMOD and
then leaves the rest up to PositionedAudio::InitFmodSound.
The only additional work that EarconSource has to do, is that it has to
monitor for the playback completing so that it can mark the Earcon as
EOF.
The other minor change is to start all audio paused, then set its
position, and then take it out of pause to start playing.