StreetPreview waits for updated TileGrid before continuing
Prior to this change, trying to use StreetPreview to a far away location
was broken and would jump to the edge of the current TileGrid instead. This
change causes the UI to wait for the new TileGrid to be uploaded before
jumping to the nearest road and providing a choice for the user.