Monitor app state and use it in heading logic along with isFlat()
This change is to try and match the iOS heading logic. The phone heading
is used if the phone is held flat or if the app is running in the
foreground (Soundscape UI is showing on the screen). Otherwise, the phone
heading is set to null and the callouts need to deal with that.