Add presentationHeading support as per iOS app
UserGeometry had already been changed to add support for multiple sources of
headings - travel direction, phone and in future head tracking. However, the
3D audio positioning was still always being done based on the phone heading.
Most of the change here is changing that to use the presentationHeading which
when moving is the travel heading and is the phone heading the rest of the
time.
This required moving the audio engine update out of the AndroidDirectionProvider
and into a combined flow handler in GeoEngine. That is now responsible for
updating the AudioEngine with the heading.
Still to do is to figure out what the AudioEngine should be doing if the
heading is null. We should at least move the audio beacons to 'dim' mode.