Import route now goes via Add Route Screen and some UI tweaking
Importing a route via an intent now goes via the add route screen. This
passes JSON describing the route from the intent into the screen which
then displays it and allows adding it to the database.

There are also numerous amateur tweaks to improve the click sites within
the markers/routes UI. Plenty still to do.