Fixup lint complaint about non-positional format string in resources
First warning was:

Multiple substitutions specified in non-positional format of string
resource string/search_results_found_first_result

The problem existed in all languages. Fix is to replace the two %s
with %1$s and %2$s.

The second warning was:
Multiple substitutions specified in non-positional format of string
resource string/callouts_hardware_low_battery

In this case, the issue was a lack of escaping of a % symbol i.e. %%