Use new theme and ensure the correct colors are used
The initial requirement for this app was that it worked the same as the
iOS app. This meant matching colors which resulted in a bit of a mess of
different color usage (at last in part my fault!). This change uses a newly
generated set of Material3 theme colors using:

   https://material-foundation.github.io/material-theme-builder/

Then the aim was to only use colors from the theme and to always use the
correct pairings e.g. onPrimary on top of primary, and onSurface on top of
surface. The screenshot test code was extended to enable a visual check of
this. In PreviewTest setting testTheme to true makes the screenshots use a
custom test theme where the pairs of colours (Xxxx and onXxxx) are the same
Color. If any of the screenshots have visible text or a visible icon, then
there's a mismatch in the theme color usage that could cause a problem with
some themes.
With these changes in place, it should now be straightforward to create and
use a new theme, or to allow custom ones. Note that currently, the app will
always use the medium contrast Light theme.

Obviously this does change how the app looks from a color perspective, but
I'm hoping that it's seen as an improvement for usability and accessibility.