Add more talkback hints and refactor IconWithTextButton
This commit has way too many different changes in it, but they are all
linked.

- There's a new talkbackHint Modifier which makes setting of a hint easier
to do.
- As part of that I added a setting to disable hints. However, all that did
was to revert to the Android default of "Doublt tap to activate" which wasn't
useful. However, I've left the code that uses the setting in place just in case
it's useful later.
- IconWithTextButton was duplicated and in general has way too many parameters.
When adding the hint, I got rid of the duplication and rationalised the call sites.
- Whilst adding a setting for Hint control I also added two to control the Theme
that's in use. Light\Dark and some contrast control. The contrast control is less
useful, but it's useful code if we come to extend the Theming.