Reduce false readings of travelHeading
This needs more work! With my phone sitting in my office locked and on its
side a playing beacon will un-dim every so often when a Location comes in
that suggests that it has a speed and a heading. Can we Kalman filter this
away? I would have expected the FusedProvider to know that there's no movement
taking place as it's looking at the accelerometers as well as the GPS, but
perhaps I'm missing something?