Improve audio beacon dimming (#422)
* Fix dimming audio so that it only affects beacons

This change creates two FMOD ChannelGroups, one for beacons and one for
speech and earcons. The audio dimming due to unknown heading is only done
for beacons.

* Reduce false readings of travelHeading

This needs more work! With my phone sitting in my office locked and on its
side a playing beacon will un-dim every so often when a Location comes in
that suggests that it has a speed and a heading. Can we Kalman filter this
away? I would have expected the FusedProvider to know that there's no movement
taking place as it's looking at the accelerometers as well as the GPS, but
perhaps I'm missing something?

* Discard polygons from MVT where all points are outwith the tile

This is a simple de-duplication which discards polygons where all of the
points are outwith the tile. PoiTest updated to reflect the fewer POIs.