Add Places Nearby hierarchy to Add Waypoints screen (#424)
* Add instructions document for initial testing

* Fix one time crash in emulator on textToSpeech startup

In updateSpeech, textToSpeech.voices was null. It was called from onInit and so
the TextToSpeech had been initialized successfully. Protect the use of voice by
checking for null.

* Remove de-duplication in PlacesNearby now that polygons are merged

There should no longer be any duplication, so we don't need a special
case to handle it.

* Move PlacesNearby viewmodel logic into separate file for sharing

We need to implement PlacesNearby in AddWaypoints, so share the code
from the viewmodel.

* Add PlacesNearby to Add Waypoints screen

I'm not sure about how tidy this is, but it does seem to do what we want.
The PlacesNearby hierarchy now appears in AddWaypoints, and selecting any
POI will pop up the SaveMarker screen. Saving the marker will save the
marker and automatically add it to the route.

* Fix URL to be a link in testing doc and add requirements section