Remove ActivityRecognition code as it was unreliable
The Android ActivityRecognition API was very delayed in detecting an
activity and appeared unreliable. It also required us to meet different
app requirements as using that API makes the app a "Health" app.
This change removes its use along with the Otto library which we were
using to pass around its events. In its place we just check if the
speed is over 5m/s. Speed will require some smoothing as it's
currently a bit jumpy and we don't use the speed error provided by
the location provider.