Start refactor of getCurrentUserGeometry
Some of the user geometry creation logic was only happening in the geometry
created for the audio engine. This change shares that code so that we have
a common set of logic. In future, the aim is to have a single collect working
on the Location/Orientation which would result in a flow for UserGeometry
which could be consumed elsewhere in the system. Currently, the Location flow
makes it all the way into the UI.