Fix crash when adding marker from Places Nearby
With no markers in the database, creating a route and adding a marker
from within AddWaypoints was causing a crash. The problem was that the
code in createAndAddMarker was assuming that createMarker had update the
objectId in locationDescription, but that update was happening on a
different thread.
The change means that the markers from the database are constantly
monitored. If a new marker appears during the lifetime of the ViewModel
it can only have been added via Add Waypoints and so it's added to the
route. This makes the update asynchronous which is as it should be.