Move markerTree into GridState and use it in Auto Callouts
The markerTree which is updated from the Realm database is now inside the
GridState class. This means that it can be easily used by callouts. Markers
in the tree now have their category set to "marker" which allows them to be
identified during callouts. The auto callout adds Markers into it's search
which should behave similarly to the iOS app.