Use country as well as language when matching locale
The code to show the default language in the onboarding screen was only
using the phone language and not the country. As a result on UK English
phones it was showing English (US) as the selected language. This change
adds the country into the mix. If the country isn't supported, the language
selected will be the first in the list i.e. French from France, non-UK English
and Portuguese from Portugal.
If the system language isn't supported, then indexOfLanguageMatchingDeviceLanguage
returns -1 which forces the user to select a language in the UI.