Fix StreetPreview behaviour to work with NearestRoad changes
The NearestRoadFilter uses the direction of travel along with the location
to determine the nearest road. If we provide the direction of travel to the
StaticLocationProvider from the StreetPreview go function then we can indicate
which road we arrived at the intersection on. The direction of travel is
always the heading of the road that was travelled along.