Remove two separate approaches to complex intersections
We have to pick an approach, so just do our best effort. A priority integer is now
created for each intersection. The more uniquely named roads other than the one that
we are on, then the higher the priority. We pick the nearest intersection with at least
one other named road, and if there's isn't one, then we pick the highest priority that
we do have.