Move to purely testing on MVT generated data
Many of the tests were still running on Soundscape backend server data. This change
moves to MVT so that we can test our map as well as our code. Whenever a new map is
generated, we run ./update.sh to get the latest tiles and check that all of the
tests still run. It may be that some new mapping data means that the assert values
change, but that's okay.