Start of adding better topology data structures
We want road segments connected by intersections that we can traverse more easily.
This is the start of that work. The WayGenerator class creates road/path features
(Way) which span only two intersections. The Intersection class contains a list of
all of the Ways that meet at that intersection. Each Way also contains the
intersection at the other end of the Way. This should make traversal simple.
Currently, these are all just turned into FeatureCollections and returned so that
they pretty much map what we currently have, other than the ways being exploded.