Create FileGridState for unit tests so as to mimic app behaviour
FileGridState is like ProtomapsGridState except that it gets the tile data
from local unit test files instead of from the server. This makes it much
easier to test what happens when the grid changes whilst changing location.
testMovingGrid has been added which takes the GPS track from nmea.csv and
replays it, moving the grid as it goes.
If we add debug recording of GPS tracks, then this is an ideal way to debug
as it can be done within the unit test framework.