Dijkstra routing using Way/Intersection graph
This is basically the same as Adam's original code, but traversing the new Way and
Intersection graph. A couple of changes to the algorithm:

 * It stops looking further from nodes which are farther away than a passed in
   maximum.
 * Once it's found a path to the end Intersection, it stops looking at nodes which
   exceed that can only exceed the shortest path (this is possibly an A* search
   rather than Djikstra uses the direction distance to the end from an Intersection
   as the best possible distance from that node).

It's around 30x faster than the original proof of concept alorithm, taking just a
few milliseconds in my testing, though it's obviously highly dependent on the data
being used. This should mean that we can use it as part of the MapMatchFilter.