LocationDetails start beacon now creates a temporary Route
By creating a temporary Route for the RoutePlayer that contains a single
Waypoint, we can re-use the UI that is used for Route control. However,
because there's no other information about the Route the Info icon is
altered to be a Stop icon instead. This allows easy stopping of the
beacon.