Confection improvements
The most complex change deals with the case when the current map matched way is a sidewalk.
It moves intersection description to be from the perspective of the road associated with the
sidewalk. It also ignores any paths that ONLY link the path with the associated road. The
downside of the road centric view is that paths coming off the sidewalk don't get called out,
though this does cut down on endless sidewalk junction call outs and so the upsides outweigh
it.

To aid testing, I've added text output of the callouts from the `testMovingGrid` unit test.