Improve map math behaviour with sidewalks
MapMatchFilter didn't allow switching to a different mapped way if it wasn't
likely based on Dijkstra determining the distance travelled. This change loosens
that so that if the old and new ways are related due to being sidewalks for the
same way, then they can be jumped between at will.