Switch to using `last buffer` style GPX recording
If the "Enable recording of travel" setting is switched on, then each GPS location is
saved to a buffer. The buffer stores up to the last hour of locations and those can be
shared to another application via the "Share recording of travel" option in the menu.
No location is shared otherwise, so it's purely a user decision. The aim of the shared
GPX file is to aid debugging. With the option enabled, a user can share their movement
if there's some behaviour that they didn't expect.