Use Dijkstra to find way into intersection during callout
When there are lots of intersections close together, we need to be able to follow our
current Way into an intersection in order to describe it correctly. Prior to this
change, intersection description was skipped if the current matched Way wasn't a
member of it, but now we use the Dijkstra search to follow our Way into the
intersection and the callouts are restored.