Fix MapLibreContainer lifetime issue
The code being used to manage the map UI lifetime wasn't working correctly.
The ON_DESTROY event was never being received and so the MapViews were
not being destroyed. In addition, the removal of the long click handler
was resulting in a warning.
The changes mean that the disposal of the MapView does now always happen,
and also the long click handler is removed correctly when the disposal takes
place.