MapMatchFilter and intersection callout fixes based on user testing
Issue #506 had missing intersection callouts which was due to a miscalculation of
the angle between two headings. It should have been calculating the angle between
the lines assuming them to be bidirectional rather than just the headings.

The MapMatchFilter has been tweaked to better deal with short spells where a
route is deviated from.