Improve name confection for un-named Ways
The previous code had a few drawbacks which this change fixes:

* Ways weren't followed to their ends, so when they cross a bridge, go
down steps or cross a tile boundary the name confection was happening at
that point instead of at the furthest known extent. The code now uses
FollowWays to get a better start and end location for confecting the names.

* Ways that cross into polygons such as parks weren't being well described
because both ends were being marked as in the park. The code now takes more
care when a Way starts and ends in/out of a Polygon to ensure that the
later feature tree searches don't end up masking that fact. It can also use
features that are within large polygons such as parks, so that paths within
parks can be named as "to cafe" etc.

* The POI used to confect were only Landmarks, but now it's landmarks and
places which makes more sense e.g. service to "Kwik Fit"

* Entrances are added as high priority POI. It's relatively few buildings
that have their entrances mapped, but it's useful to use them in name
confection as they have much more precision than large POI polygons like
supermarkets etc.

One side issue was ensuring that rulers were passed into FeatureTree for
searching which resulted in a lot of API changes in unit tests.