Initial attempt at dynamic accessible map coloring
I think that this is definitely better than the default map style,
though I suspect that there's much more work to improve it. By default
the original OSM liberty style is used, but the new map style can be
enabled by going to the Settings and turning on "Enable accessible map
styling". With that on, the accessible style will be used which
currently switches between light and dark based on the theme. This will
allow visually impaired testers to provide some feedback.