Implement full screen map mode
Whenever there is a map on screen, there's now a Floating Action Button
which will make the map full screen. Once displaying in full screen mode,
the FAB changes to be an 'exit full screen' icon.

The addition of this behaviour made #398 far more obvious, and so that
too was fixed in this commit. It was a race between map rendering and
adding of beacon and route markers.