Add POI callouts to unit test GPX replay
Two parts to this:

1. Setting the enabledCategories for the main grid state in getGridStateForLocation.
Without this, the PLACES_AND_LANDMARKS_KEY FeatureTree remains empty.
2. Add a timestamp into UserGeometry and use that as the current time. For the
regular app, this is just the system time in milliseconds, but for unit tests we can
use the value from the GPX file. This gives us the correct timing for tracking
callouts whereas before we would always repeat callouts.