Fix up route and marker editing so that it works correctly
Several things going on here:

1. The orderId is now used as the unique identifier for re-ordering routes.
This means that in theory we could have the same marker in a route twice which
might be useful. The main reason for the change is that until markers have been
added to the database they don't have a database id.

2. Route editing was getting confused as it was changing the uiState underneath
the add waypoint list. I couldn't reproduce issues prior to the Room DB changes,
but the problem isn't new.

3. Using intents to add routes wasn't working more than once. Resetting the
navigator destination resolves this.

4. Disallow creating a duplicate marker at the same location when importing from
a GPX file. This should make it easier to import a load of routes that share
markers.