Privacy policy and terms and conditions
These are based on the iOS policies from https://scottishtecharmy.freshdesk.com/support/solutions

Currently the only data that makes it off the device is:

1. The precise location at the time of a crash.
2. The GPX track of the location over the last hour when the user chooses to
enable recording from the Settings menu and then Shares it from the main menu.
3. The current location when the user selects Share from the Location Details
page.

Both of the latter two allow the user to select another app e.g. email, messaging
through which to Share the data. It doesn't go to any of our own servers, so only
the first is 'collected' and 'shared'.