Add Scaffold to Drawer menu and version in its bottom bar
The current version wasn't easily visible, so put it in the bottom bar of the drawer
menu where it's very easy to find. Adding the scaffold means that we can add far more
menu entries and support different screen formats and the user will be able to scroll
to get to them.