Add test to generate screenshots for PlayStore
The reason for having these as instrumented tests is that we want them to
include the rendered map. However, because that requires the GitHub secret
containing the TILE_PROVIDER it can't be done from any old build. We can do
it from a release, so do it there and upload the artifacts to that release.
We can have both screenshots for the PlayStore, and ones for including in the
jekyll docs build.

We don't want the time to appear on the screenshots as that would result in
endlessly changing screenshots, so trim them with imagemagick before committing.
Any changes will result in the docs being rebuilt and using the new images.