Initial audio focus work
There are two things happening here:

1. The addition of calls to try and attain Android Audio Focus.
2. Moving the history/filtering additions for callbacks to a point where
we're confident that the audio is actually being played out. If we don't
have audio focus, then we don't want to add un-played callbacks to the
callout history as that would stop them being played back once Audio Focus
is restored.

The app will always try and gain audio focus if there's a beacon playing and
it's visible on the screen. It also will try and gain focus whenever there's
a speech callout. Something like listening to Spotify is generally unlikely
with users but it will work well so long as there are no audio beacons. If
there's an audio beacon, then Spotify will end up mostly ducked which although
correct isn't great for listening to music, but the beacon is prioritized.

Contention with Talkback is generally better. When Talkback is speaking, the
Soundscape content is ducked which is what you'd expect.