Drive audio focus abandonment from C++ AudioEngine
Instead of trying to figure out whether audio is playing from the service,
add a callback from the AudioEngine to tell us when no more audio is playing.
When called we can abandon the audio focus knowing that there's nothing
playing.
The callback is called:

1. When the size of the audio beacons list goes down to zero during UpdateGeometry.
This happens when text to speech audio ends.
2. When a beacon is destroyed via a call to destroyNativeBeacon. This happens
when an audio beacon is stopped from the app.