Add Auto mode for light/dark color theme selection
We want to start the app with the system set theme as dark if that's
what the phone is set to. This is done by adding an "Auto" option which
can be over-ridden with Light or Dark.

Auto has also been added to the measurement units where the Locale is used
to determine whether to use metric or imperial units. This can also be
over-ridden with Metric or Imperial.