Adjust light and dark themes to be more light and dark
This switches to using surfaceContainer throughout. It does result
in a very limited range of colors, but the contrast is good. Need
some user feedback.
It also ensures that the onboarding screens honour the system dark
theme setting. An we also re-enable screenshots which had been
broken by an update to the library. Not all screenshots are working,
so there's a bug filed on that.