Move TTS functionality into its own class TtsEngine to enable selecting
The TextToSpeech code was tightly intertwined with the AudioEngine code which
made switching engines and voices harder than it needed to be. This change:

1. Moves all of the TextToSpeech code into the TtsEngine class
2. Adds TTS engine selection from Settings as well as voices
3. Fixes an issue with the Settings list items with large text
4. Fixes a small issue with the TTS playback introduced by the sample rate work

Now the user can switch TTS engines and select voices from each and it should cope
with all audio formats and sample rates. The only thing I haven't got working is
updating the Settings summary when the engine changes.