Help users when missing location permissions
If a user denies location permissions the only way to get them is via
app settings. This change adds a button to go there where the map normally
is on the home screen. This aims to make it clear why the app isn't
working fully.
The Soundscape service is started either in onResume() or after checking the
permissions. This means that the app now deals with the Location setting being
enabled whilst the app is running. However, it doesn't yet correctly deal with
the permission being disabled whilst the app is running.