Improvements on permission changes
The previous permissions changes broke Sleep mode if the permissions
were changed whilst in sleep. They also didn't offer permissions as often
as they could. This change monitors the permissions during onResume and
looks for them changing. That is then used in the GUI for displaying the
permissions help button. This prevents a bit of GUI flicker when going in
and out of Sleep.