Fix photon search language selection
Our photon search server currently only supports English, French, German and
default languages. Passing in any other language returns an error.
This change adds a new SearchLanguage settings which defaults to "auto". It
allows picking one of the supported languages to use with the search engine.
This is so that someone using a language such as Portuguese in Japan can
get the English search results rather than the Japanese ones.