Catch exception when starting foreground service
A couple of changes here.

Firstly, catch the ForegroundServiceStartNotAllowedException
and log it rather than crash. In that case don't mark the service as
running so that it will try and start again the next time.

Secondly, trigger the service notification immediately rather than
waiting for 10 seconds which was just confusing.