Auto callout tweaks
1. Markers are now only called out if they are within the field of view - a 90 degree
quadrant centered on straight ahead. This stops markers from being called out when the
user has passed them.

2. In Street Preview 'jump to next intersection' mode there's only one useful location
update per jump. Because there's always an intersection, this meant that no POI were ever
called out. This change means that any nearby POI are always tagged on to the end of the
intersection callout when in Street Preview mode.