Reinstate Kalman filtering for all location work other than map matching
When map matching was added, Kalman filtering on the incoming location was
removed. This was a mistake as that resulted in the raw location complete
with GPS errors being used in a lot of the code. This change reinstates
Kalman filtering for the location in UserGeometry and has a special case
for using the raw location for map matching.