More crashlytics logging aimed and changes aimed at foreground start
If startForeground fails we now start the geoEngine anyway. And then in the
next onResume we try startForeground again. There's also some debug code to
make it 'fail' the first time to check the error paths. If there's no
onResume within 10 seconds then the app will still shutdown, but not due to
a crash. I think that's as good as we can do for now.
The Crashlytics logging should allow us to check that any failures are a
due to this behaviour.